Split types from utils and move into zephyrus-search/tree folder
